+++ Reach/test/utopia/reach/test/DropDownTest.scala
@@ -30,7 +30,9 @@ object DropDownTest extends App
 	val arrowImage = Image.readFrom("Reflection/test-images/arrow-back-48dp.png")
 	arrowImage.logFailure
 	val expandIcon = arrowImage.map { i => new SingleColorIcon(i.transformedWith(Matrix2D.quarterRotationCounterClockwise)) }
+		.getOrElse(SingleColorIcon.empty)
 	val shrinkIcon = arrowImage.map { i => new SingleColorIcon(i.transformedWith(Matrix2D.quarterRotationClockwise)) }
+		.getOrElse(SingleColorIcon.empty)
 	
 	val items = Map("Fruits" -> Vector("Apple", "Banana", "Kiwi"), "Minerals" -> Vector("Diamond", "Ruby", "Sapphire"))
 	
@@ -46,15 +48,15 @@ object DropDownTest extends App
 						selectedItemPointer.addListener { e => println(e) }
 						
 						Vector(
-							ddF.simple(Fixed(items.keys.toVector.sorted), selectedCategoryPointer, expandIcon.toOption,
-								shrinkIcon.toOption, fieldNamePointer = Fixed("Category"),
+							ddF.simple(Fixed(items.keys.toVector.sorted), selectedCategoryPointer, expandIcon,
+								shrinkIcon, fieldNamePointer = Fixed("Category"),
 								promptPointer = Fixed("Select One")),
 							ddF.simple(
 								selectedCategoryPointer.map {
 									case Some(category) => items(category)
 									case None => Vector()
 								},
-								selectedItemPointer, expandIcon.toOption, shrinkIcon.toOption,
+								selectedItemPointer, expandIcon, shrinkIcon,
 								fieldNamePointer = selectedCategoryPointer.map {
 									case Some(category) => category
 									case None => "Item"